+++ src/FsCodec.NewtonsoftJson/Options.fs
@@ -17,13 +17,13 @@ type Options private () =
     /// Creates a default set of serializer settings used by Json serialization. When used with no args, same as JsonSerializerSettings.CreateDefault()
     static member CreateDefault
         (   [<Optional; ParamArray>] converters : JsonConverter[],
-            /// Use multi-line, indented formatting when serializing JSON; defaults to false.
+            // Use multi-line, indented formatting when serializing JSON; defaults to false.
             [<Optional; DefaultParameterValue(null)>] ?indent : bool,
-            /// Render idiomatic camelCase for PascalCase items by using `CamelCasePropertyNamesContractResolver`. Defaults to false.
+            // Render idiomatic camelCase for PascalCase items by using `CamelCasePropertyNamesContractResolver`. Defaults to false.
             [<Optional; DefaultParameterValue(null)>] ?camelCase : bool,
-            /// Ignore null values in input data; defaults to false.
+            // Ignore null values in input data; defaults to false.
             [<Optional; DefaultParameterValue(null)>] ?ignoreNulls : bool,
-            /// Error on missing values (as opposed to letting them just be default-initialized); defaults to false.
+            // Error on missing values (as opposed to letting them just be default-initialized); defaults to false.
             [<Optional; DefaultParameterValue(null)>] ?errorOnMissing : bool) =
 
         let indent = defaultArg indent false
@@ -48,16 +48,16 @@ type Options private () =
     /// - Always prepends an OptionConverter() to any converters supplied
     /// - everything else is as per CreateDefault:- i.e. emit nulls instead of omitting fields etc
     static member Create
-        (   /// List of converters to apply. An implicit OptionConverter() will be prepended and/or be used as a default
+        (   // List of converters to apply. An implicit OptionConverter() will be prepended and/or be used as a default
             [<Optional; ParamArray>] converters : JsonConverter[],
-            /// Use multi-line, indented formatting when serializing JSON; defaults to false.
+            // Use multi-line, indented formatting when serializing JSON; defaults to false.
             [<Optional; DefaultParameterValue(null)>] ?indent : bool,
-            /// Render idiomatic camelCase for PascalCase items by using `CamelCasePropertyNamesContractResolver`.
-            ///  Defaults to false on basis that you'll use record and tuple field names that are camelCase (and hence not `CLSCompliant`).
+            // Render idiomatic camelCase for PascalCase items by using `CamelCasePropertyNamesContractResolver`.
+            //  Defaults to false on basis that you'll use record and tuple field names that are camelCase (and hence not `CLSCompliant`).
             [<Optional; DefaultParameterValue(null)>] ?camelCase : bool,
-            /// Ignore null values in input data; defaults to `false`.
+            // Ignore null values in input data; defaults to `false`.
             [<Optional; DefaultParameterValue(null)>] ?ignoreNulls : bool,
-            /// Error on missing values (as opposed to letting them just be default-initialized); defaults to false
+            // Error on missing values (as opposed to letting them just be default-initialized); defaults to false
             [<Optional; DefaultParameterValue(null)>] ?errorOnMissing : bool) =
 
         Options.CreateDefault(